Avoid cinder calls, when cinder is unavailable
When a volume service is not available, then a cinder client can not
be created. This patch skips the calls, so that dashboard doesn't
break any more.

This is a backport to Folsom release, as this turned out to be a major
issue.
Fixes bug 1084137

Change-Id: I8f2f8b0b131b4bb5319d74f6da48671f146d7e00
Fixes bug 886284
Upstream change https://review.openstack.org/#/c/18303/
Reviewed-on: https://code.engineering.redhat.com/gerrit/2747
Reviewed-by: Alan Pevec <apevec@redhat.com>
Tested-by: Alan Pevec <apevec@redhat.com>