Add support for row actions to detail pages
Add support for performing actions on an object from the object's
detail page. Actions are displayed in the page header, in a
row of buttons. To use this newly added capability, `row=True`
must be passed as a parameter to the render_row_actions method,
when called from a DetailView.

Change-Id: I4fd96433f9514242c21d6c47f79bf83f3a5e4fc0
Implements: blueprint detail-pages-ia